Move EntryID and EntryName to parsec.api.data
