+++ parsec/api/data/message.py
@@ -4,7 +4,7 @@ from pendulum import Pendulum
 
 from parsec.crypto import SecretKey
 from parsec.serde import fields, post_load, OneOfSchema
-from parsec.core.types.base import EntryID, EntryIDField
+from parsec.api.data.entry import EntryID, EntryIDField
 from parsec.api.data.base import BaseAPISignedData, BaseSignedDataSchema
