+++ parsec/core/types/local_device.py
@@ -6,8 +6,7 @@ from hashlib import sha256
 from parsec.crypto import SecretKey, PrivateKey, SigningKey
 from parsec.serde import fields, post_load
 from parsec.api.protocol import DeviceID, OrganizationID, DeviceIDField
-from parsec.api.data import BaseLocalData, BaseSchema
-from parsec.core.types.base import EntryID, EntryIDField
+from parsec.api.data import BaseLocalData, BaseSchema, EntryID, EntryIDField
 from parsec.core.types.backend_address import BackendOrganizationAddr, BackendOrganizationAddrField
